bcm43455c0: first try to handle amsdus in monitor mode. to test it, comment the following lines: __attribute__((at(0x1B6B02, , CHIP_VER_BCM43455c0, FW_VER_7_45_189))) BPatch(wlc_monitor_amsdu_patch, 0x1B6B1E);
The patch will check the size of a frame before pushing the frame beginnen to add a header. If there is not enough space an error will be printed and the frame will be ignored. To have enough space for AMSDU frames, the size when allocating a new duplicated buffer in the wlc_monitor_amsdu function is increased. This could result in a buffer size that is too large to be supported which we might need to check next.

Patch is currently untested.